add R help pages for package and `query` function
There was a request to add HTTP headers such as basic auth. The
implementation did actually support this feature, but it was not
documented as well as it should have been. This change simply adds the
neccesary documentation to make it easier for users to understand how to
add additional configuration to the http client.